+++ kvm/playground/cpuid.c
@@ -18,7 +18,7 @@
 #include <errno.h>
 #include <stdlib.h>
 
-#include "common.h"
+#include "kvm_helpers.h"
 
 extern uint8_t start_of_code;
 extern uint8_t end_of_code;