+++ kvm/playground/kvm_helpers.c
@@ -9,7 +9,7 @@
 #include <linux/kvm.h>
 #include <stdio.h>
 
-#include "common.h"
+#include "kvm_helpers.h"
 
 int open_dev_kvm(void)
 {
@@ -20,7 +20,8 @@ int open_dev_kvm(void)
 	}
 	int ret = ioctl(kvm, KVM_GET_API_VERSION, NULL);
 	if (ret != KVM_API_VERSION) {
-		fprintf(stderr, "KVM_GET_API_VERSION expoected 12 but got %d.", ret);
+		fprintf(stderr, "KVM_GET_API_VERSION expoected 12 but got %d.",
+			ret);
 		exit(1);
 	}
 
@@ -31,7 +32,8 @@ int run_vcpu(int vcpufd)
 {
 	int ret;
 	if ((ret = ioctl(vcpufd, KVM_RUN, NULL)) == -1) {
-		printf("KVM_RUN was unable to start the VM. Error code: %d\n", ret);
+		printf("KVM_RUN was unable to start the VM. Error code: %d\n",
+		       ret);
 		exit(1);
 	}
 	return ret;
@@ -41,7 +43,9 @@ void check_cap_user_memory(int kvmfd)
 {
 	int ret = ioctl(kvmfd, KVM_CHECK_EXTENSION, KVM_CAP_USER_MEMORY);
 	if (ret == -1) {
-		fprintf(stderr, "KVM_CAP_USER_MEM not available. Error code: %d\n", ret);
+		fprintf(stderr,
+			"KVM_CAP_USER_MEM not available. Error code: %d\n",
+			ret);
 		exit(1);
 	}
 }
@@ -54,7 +58,9 @@ void check_cap_ext_cpuid(int kvmfd)
 {
 	int ret = ioctl(kvmfd, KVM_CHECK_EXTENSION, KVM_CAP_EXT_CPUID);
 	if (ret == -1) {
-		fprintf(stderr, "KVM_CAP_EXT_CPUID not available. Error code: %d\n", ret);
+		fprintf(stderr,
+			"KVM_CAP_EXT_CPUID not available. Error code: %d\n",
+			ret);
 		exit(1);
 	}
 }
@@ -67,7 +73,9 @@ void check_cap_ext_emul_cpuid(int kvmfd)
 {
 	int ret = ioctl(kvmfd, KVM_CHECK_EXTENSION, KVM_CAP_EXT_EMUL_CPUID);
 	if (ret == -1) {
-		fprintf(stderr, "KVM_CAP_EXT_EMUL_CPUID not available. Error code: %d\n", ret);
+		fprintf(stderr,
+			"KVM_CAP_EXT_EMUL_CPUID not available. Error code: %d\n",
+			ret);
 		exit(1);
 	}
 }
@@ -76,7 +84,9 @@ void check_cap_get_msr_features(int kvm)
 {
 	int ret = ioctl(kvm, KVM_CHECK_EXTENSION, KVM_CAP_GET_MSR_FEATURES);
 	if (ret == -1) {
-		fprintf(stderr, "KVM_CAP_GET_MSR_FEATURES not available. Error code: %d\n", ret);
+		fprintf(stderr,
+			"KVM_CAP_GET_MSR_FEATURES not available. Error code: %d\n",
+			ret);
 		exit(1);
 	}
 }
@@ -95,7 +105,9 @@ int create_vcpu(int vmfd)
 {
 	int vcpufd = ioctl(vmfd, KVM_CREATE_VCPU, (unsigned long)0);
 	if (vcpufd == -1) {
-		fprintf(stderr, "Could not create VCPU for VM %d. Error code: %d", vmfd, vcpufd);
+		fprintf(stderr,
+			"Could not create VCPU for VM %d. Error code: %d", vmfd,
+			vcpufd);
 		exit(1);
 	}
 	return vcpufd;
@@ -125,7 +137,8 @@ void *mmap_user_memory_region(int vmfd, size_t size, unsigned long gpa,
 
 	ret = ioctl(vmfd, KVM_SET_USER_MEMORY_REGION, &region);
 	if (ret == -1) {
-		fprintf(stderr, "Could not set guest memory. Error code: %d\n", ret);
+		fprintf(stderr, "Could not set guest memory. Error code: %d\n",
+			ret);
 		exit(1);
 	}
 
@@ -139,7 +152,7 @@ struct kvm_run *mmap_kvm_run(int kvmfd, int vcpufd)
 
 	mmap_size = ioctl(kvmfd, KVM_GET_VCPU_MMAP_SIZE, NULL);
 	run = (struct kvm_run *)mmap(NULL, mmap_size, PROT_READ | PROT_WRITE,
-					MAP_SHARED, vcpufd, 0);
+				     MAP_SHARED, vcpufd, 0);
 
 	return run;
 }
@@ -148,7 +161,8 @@ void get_sregs(int vcpufd, struct kvm_sregs *sregs)
 {
 	int ret = ioctl(vcpufd, KVM_GET_SREGS, sregs);
 	if (ret == -1) {
-		printf("KVM_GET_SREGS failed to read special registers. Exit code: %d\n", ret);
+		printf("KVM_GET_SREGS failed to read special registers. Exit code: %d\n",
+		       ret);
 		exit(1);
 	}
 }
@@ -157,7 +171,8 @@ void set_sregs(int vcpufd, struct kvm_sregs *sregs)
 {
 	int ret = ioctl(vcpufd, KVM_SET_SREGS, sregs);
 	if (ret == -1) {
-		printf("KVM_SET_SREGS failed to update special registers. Exit code: %d\n", ret);
+		printf("KVM_SET_SREGS failed to update special registers. Exit code: %d\n",
+		       ret);
 		exit(1);
 	}
 }
@@ -165,26 +180,17 @@ void set_sregs(int vcpufd, struct kvm_sregs *sregs)
 void dump_kvm_sregs(struct kvm_sregs *sregs)
 {
 #if defined(__x86_64__)
-#define segment(name)	\
-	printf("%-5s %#016llx %#08x %#08x %-4d\n", \
-		#name, \
-		sregs->name.base, \
-		sregs->name.limit, \
-		sregs->name.selector, \
-		sregs->name.type \
-	);
-#define dtable(name)	\
-	printf("%-5s %#016llx %#08x\n", \
-		#name, \
-		sregs->name.base, \
-		sregs->name.limit \
-	);
-#define pu64(name)	\
-	printf("%-5s %#016llx\n", #name, sregs->name);
+#define segment(name)                                                       \
+	printf("%-5s %#016llx %#08x %#08x %-4d\n", #name, sregs->name.base, \
+	       sregs->name.limit, sregs->name.selector, sregs->name.type);
+#define dtable(name)                                             \
+	printf("%-5s %#016llx %#08x\n", #name, sregs->name.base, \
+	       sregs->name.limit);
+#define pu64(name) printf("%-5s %#016llx\n", #name, sregs->name);
 
 	printf("Dump kvm_sregs\n");
-	printf("%-5s %-16s %-8s %-8s %-4s\n",
-		"NAME", "Base", "Limit", "Selector", "Type");
+	printf("%-5s %-16s %-8s %-8s %-4s\n", "NAME", "Base", "Limit",
+	       "Selector", "Type");
 	segment(cs);
 	segment(ds);
 	segment(es);
@@ -208,7 +214,7 @@ void dump_kvm_sregs(struct kvm_sregs *sregs)
 #undef dtable
 #undef pu64
 #else
-# error "Unsupport arch"
+#error "Unsupport arch"
 #endif
 }
 
@@ -216,7 +222,8 @@ void get_regs(int vcpufd, struct kvm_regs *regs)
 {
 	int ret = ioctl(vcpufd, KVM_GET_REGS, regs);
 	if (ret == -1) {
-		printf("KVM_GET_REGS failed to read special registers. Exit code: %d\n", ret);
+		printf("KVM_GET_REGS failed to read special registers. Exit code: %d\n",
+		       ret);
 		exit(1);
 	}
 }
@@ -225,7 +232,8 @@ void set_regs(int vcpufd, struct kvm_regs *regs)
 {
 	int ret = ioctl(vcpufd, KVM_SET_REGS, regs);
 	if (ret == -1) {
-		printf("KVM_SET_REGS failed to update special registers. Exit code: %d\n", ret);
+		printf("KVM_SET_REGS failed to update special registers. Exit code: %d\n",
+		       ret);
 		exit(1);
 	}
 }
@@ -233,18 +241,34 @@ void set_regs(int vcpufd, struct kvm_regs *regs)
 void dump_kvm_regs(struct kvm_regs *regs)
 {
 #if defined(__x86_64__)
-#define pu64(name)	\
-	printf("%-5s %#016llx ", #name, regs->name);
+#define pu64(name) printf("%-5s %#016llx ", #name, regs->name);
 
 	printf("Dump kvm_regs\n");
-	pu64(rax); pu64(rbx); pu64(rcx); pu64(rdx); printf("\n");
-	pu64(rsi); pu64(rdi); pu64(rsp); pu64(rbp); printf("\n");
-	pu64(r8); pu64(r9); pu64(r10); pu64(r11); printf("\n");
-	pu64(r12); pu64(r13); pu64(r14); pu64(r15); printf("\n");
-	pu64(rip); pu64(rflags); printf("\n");
+	pu64(rax);
+	pu64(rbx);
+	pu64(rcx);
+	pu64(rdx);
+	printf("\n");
+	pu64(rsi);
+	pu64(rdi);
+	pu64(rsp);
+	pu64(rbp);
+	printf("\n");
+	pu64(r8);
+	pu64(r9);
+	pu64(r10);
+	pu64(r11);
+	printf("\n");
+	pu64(r12);
+	pu64(r13);
+	pu64(r14);
+	pu64(r15);
+	printf("\n");
+	pu64(rip);
+	pu64(rflags);
+	printf("\n");
 #undef pu64
 #else
-# error "Unsupport arch"
+#error "Unsupport arch"
 #endif
 }
-