+++ kvm/playground/kvm_helpers.h
@@ -16,8 +16,8 @@ void check_cap_ext_cpuid(int kvmfd);
 void check_cap_ext_emul_cpuid(int kvmfd);
 void check_cap_get_msr_features(int kvm);
 
-void* mmap_user_memory_region(int vmfd, size_t size, unsigned long gpa,
-		const void *code, size_t code_len);
+void *mmap_user_memory_region(int vmfd, size_t size, unsigned long gpa,
+			      const void *code, size_t code_len);
 struct kvm_run *mmap_kvm_run(int kvmfd, int vcpufd);
 
 void get_sregs(int vcpufd, struct kvm_sregs *sregs);
@@ -27,4 +27,3 @@ void dump_kvm_sregs(struct kvm_sregs *sregs);
 void get_regs(int vcpufd, struct kvm_regs *regs);
 void set_regs(int vcpufd, struct kvm_regs *regs);
 void dump_kvm_regs(struct kvm_regs *regs);
-