+++ kvm/playground/load-payload02.c
@@ -10,7 +10,7 @@
 #include <stdlib.h>
 #include <unistd.h>
 
-#include "common.h"
+#include "kvm_helpers.h"
 
 extern uint8_t start_of_code;
 extern uint8_t end_of_code;